requirements.txt
# Content Factory Skill Dependencies
# Python >= 3.8

# 文档生成
python-docx>=0.8.11

# 网络请求
requests>=2.28.0
urllib3>=1.26.0

# 数据解析
pyyaml>=6.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# 工具库
python-dateutil>=2.8.0
jinja2>=3.1.0

# 开发测试（可选）
pytest>=7.0.0
pytest-cov>=4.0.0
